Started Graphics
This commit is the first instance of the code that has some graphics in
it. The major updates in this commit include the creation of four new
classes: YellowFrame, YellowPanel, YellowTester, and Car. The Frame,
Panel, and Tester classes are meant to test and display the graphics
while the Car class is supposed to calculate the change in the x and y
coordinates in the graphics. With this update, I can close the issue
from a few days ago and make a new one with what needs to be done this
weekend,