Graphics Improvements
Added the car functionality so now a car is drawn on the map and a
crude representation of how it moves is completed. There is still a lot
of work that needs to be done in order to finish this. I will be
posting an issue to describe what needs to be done. With this update, a
single car will work, but the graphics don’t draw if we try to put the
car drawing code in a loop (the problem is clearly labeled with
comments and its in YellowPanel). Since this is a huge leap from the
previous commit, the important code has been explained through
comments. It should be easy to read and understand.