v1.0 Commit: Ready for Presentation
The code works for the most part. There are still some issues, but they
rarely occur. The code is ready for presentation. The changes in this
commit include: drawing the signal light, getting multiple cars to
work, and drawing a string to inform the user of a new car simulation.
There is very small chance that more work will be done on this project,
so this can be considered to be a final commit.